requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.6.0
python-multipart==0.0.9
google-generativeai>=0.7.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9